Multiple window and immersive mode time measurement telemetry. (#3430)
* Bump to AC 41.0.0.

* Adding loading time and immersive mode duration Glean telemetry.

* Adding multiple window Glean telemetry.

* Adding time measurement metrics to metrics.yaml.

* Adding multiple window telemetry tests for Glean.

+++ app/src/common/shared/org/mozilla/vrbrowser/browser/Places.kt
@@ -26,9 +26,11 @@ class Places(var context: Context) {
         val files = context.filesDir.listFiles { _, name ->
             name.matches("places\\.sqlite.*".toRegex())
         }
-        for (file in files) {
-            if (!file.delete()) {
-                Logger(LOGTAG).debug("Can't remove " + file.absolutePath)
+        if (files != null) {
+            for (file in files) {
+                if (!file.delete()) {
+                    Logger(LOGTAG).debug("Can't remove " + file.absolutePath)
+                }
             }
         }
